add SortValidationTest
